requirements.txt
fastapi
uvicorn
langgraph
pydantic
psycopg2-binary
python-dotenv
groq
sentence-transformers
faiss-cpu
numpy
